[AW] Add UMA for measuring shouldInterceptRequest usage and intercepted url schemes
Add UMA for measuring whether the request was intercepted by the embedder
of Android WebView and if so which url scheme was intercepted.

Intercepted requests are requests for which the shouldInterceptRequest()
callback returned a non-null response.

In particular we are interested in the usage of "data:" scheme url
interception, because we are considering not allowing to intercept
this type of scheme.

BUG=841556,928728,929788

Change-Id: I1063605bda171e3e75f259c1539626212d79c4d7
Reviewed-on: https://chromium-review.googlesource.com/c/1459116
Commit-Queue: Tim Volodine <timvolodine@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#631378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3778f893c76248a57cbf0cc7434c633561783de0