[CrOS MultiDevice] Create ClientAppMetadataProvider.
During CryptAuth v2 enrollment, the Chromebook must upload a
ClientAppMetadata proto to the back-end. This CL adds a
pure-virtual ClientAppMetadataProvider class in the DeviceSync service
which computes this proto.

Because some fields of this proto cannot be calculated from within the
service due to circular dependencies, the concrete implementation of
this class is implemented as ClientAppMetadataProviderService, a
KeyedService residing in //chrome.

Note that this CL also extracts common GCM constants to new files
gcm_constants.h/cc so that they can be shared with
CryptAuthGcmManagerImpl.

Bug: 899080
Change-Id: I68397f9b27395afdca4b2fb608a0dfa0a5555575
Reviewed-on: https://chromium-review.googlesource.com/c/1465178
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Jeremy Klein <jlklein@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#631903}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f06ef7340b0a075877b040fc32c05ca69dff3d28