ProfileProvider: adds a Chrome OS heap collector
The heap collector is enabled only on devices with the new version of tcmalloc
on the canary and dev channels.

The collector is enabled for all browser processes on the canary channel and
with a 20% probability for browser processes on the dev channel.

Added a new collection status value, UNABLE_TO_COLLECT, used when collection
cannot be started due to technical reasons.

BUG=b:121160411
TEST=Unit tests pass

Change-Id: I3b71c2b10cc00cda66bf845fc5ac6f255dade3c6
Reviewed-on: https://chromium-review.googlesource.com/c/1435075
Commit-Queue: Gabriel Marin <gmx@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#632679}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8c48c5453b8eed64c76a2590c4f8598dce478ab6