[Video Capture, Mac] Restart video capture service when no camera found on MacBooks
UMA logs show that on MacBooks, which are assumed to all have an internal
camera, it can happen that no camera is found. As per https://crbug.com/643068,
restarting the process may help in this situation.

With video capture code now running in a separate utility process, we have the
ability to restart it "behind the scenes".

This CL adds a feature "RetryGetVideoCaptureDeviceInfos" that implements such
retry logic involving a restart of the video capture service on Mac.

Bug: 643068
Change-Id: Ief06154eaee430e3f93a470b9da10f2ccc224a56
Reviewed-on: https://chromium-review.googlesource.com/c/1450291
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Weiyong Yao <braveyao@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Emircan Uysaler <emircan@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#633537}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 724807ac02f83607813cdeabff9639a3735276de