Toolchain with d3dcompiler_old.dll included
This switches Chromium for Windows to use a new toolchain package that
is the same as the one landed in crrev.com/c/1450414 except that it
includes both the new d3dcompiler_47.dll (from the 17763 SDK) and the
old one (from the 17134 SDK). The two versions are so that most
customers get the new version, with all of its bug fixes, while those
who don't have proper UCRT installs will fall back to the old version.
If there are enough of the latter then we can figure out a way to ship
the UCRT beside chrome.exe so that they can get the fixes as well. If
there are few enough of the latter then we can drop support for them.

This change also removes the UCRT from installs, reverting a portion of
crrev.com/c/1450414, and adds d3dcompiler_old.dll to installs.

This toolchain package was created by copying d3dcompiler_47.dll from
the win_sdk\Redist\D3D\{x64|x86} directories of the 17134 SDK to
d3dcompiler_old.dll and repackaging the directory using the --repackage
option of depot_tools\win_toolchain\package_from_installed.py.

Note that the logic in build\win\BUILD.gn for when to package up the
UCRT was incorrect (it would try to package it for ARM64 component
builds) and that is corrected as well in this change.

Angle can then use d3dcompiler_old.dll as a fallback, and record
metrics about how often the fallback is used.

Bug: 920704
Change-Id: Ida96d8b9247789a429519c50b0663f183d72c800
Reviewed-on: https://chromium-review.googlesource.com/c/1461679
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Jamie Madill <jmadill@google.com>
Reviewed-by: Greg Thompson <grt@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#634207}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6ddbac7f8b44ca9f4ef8adaf21229df9dec380a2