Add policy for Assistant hotword
The VoiceInteractionHotwordEnabled preference needs to be managed by an
administrator for G-Suite accounts. Added a policy for the preference.

In addition, updated the settings UI so that the widgets controlled by
the policy are disabled when controlled by ab administrator.

Bug: b:124113875
Change-Id: If2c3ff8a50b69e8694a48c1c67e12ccd3212e069
Reviewed-on: https://chromium-review.googlesource.com/c/1465070
Commit-Queue: Eyor Alemayehu <eyor@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#634798}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a2a2f7d5639e36bf498df2fb4a056a4ef615b1f9