CSS: retire use count :not with invalid selectors
This reverts the use counters added in
https://chromium-review.googlesource.com/c/chromium/src/+/1354751

The CSS Working Group was considering allowing :not() to
ignore selectors that are invalid or not accepted by the browser.

Use counts showed that this would affect a large percentage of pages
https://github.com/w3c/csswg-drafts/issues/3264#issuecomment-463036664
and so the decision was made to not change the behavior of :not
https://github.com/w3c/csswg-drafts/issues/3264#issuecomment-467190906

Data collection is no longer required.

BUG=568705

Change-Id: I38ddaee5339e19b70c1c13c9507a50ae27c354b0
Reviewed-on: https://chromium-review.googlesource.com/c/1487753
Commit-Queue: Emil A Eklund <eae@chromium.org>
Reviewed-by: Emil A Eklund <eae@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#635559}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b9a6186caaf4b814b297d2e3f7fe3d22c3b6eff