Migrate DataSaver Snackbar Promo to Generic Class
This CL is 1 of 4 to move the DataSaver "Chrome saved you XX data" from
a snackbar to an IPH.

This CL moves the snackbar controller class to a class
in preferences.datareduction and renames everything from "snackbar
promo" to a more generic "milestone promo".
All of the prefs and variation params were kept to allow the
configurations we already have to keep working as is, although this does
lead to some stale naming.

UMA is removed in this CL, and will be re-added in the final CL.

Bug: 878337
Change-Id: Ia7924cd741db2b6b37180d6fb93735636fbaf00d
Reviewed-on: https://chromium-review.googlesource.com/c/1487196
Reviewed-by: Theresa <twellington@chromium.org>
Reviewed-by: Doug Arnett <dougarnett@chromium.org>
Commit-Queue: Robert Ogden <robertogden@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#636695}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f524a658ae437e58eb2f2ebf5b4a8a0c9c266b6