Add basic UMA to inline updates
Add basic UMA to get a good understanding of how Play and Chrome are
interacting.  This is set to expire in a few months, when we are sure we
will have a good understanding of Play expectations and behavior in the
wild.

Bug: 934345
Change-Id: Ib70d661c8034acb3fe480620cc7664db29a88060
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1481856
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: David Trainor <dtrainor@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#638765}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fae90fcac158ee34b59c463c4dd817bfb18d1b2e