Redesign Clank context menu for better usability
This CL contains some code to display the "revamped" context menu as
a ListView with options that are divided into groups. As the first
item in the list, we have a header which contains a thumbnail,
title, and link.

Bug: 655359
Change-Id: Ib6a75c5b4b79b260857a00202b9782962d5cf271
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1497431
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Theresa <twellington@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#639536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3bafe27bf185d377b970415ba7037a8563a67b56