Add stub implementation of InstallableInkDrop
This adds a non-functional skeleton of InstallableInkDrop. Additionally,
it adds a feature flag for using InstallableInkDrop in place of InkDropImpl.
Only ToolbarButton support for the flag is added (for now).

The portions of the InkDrop interface not required for ToolbarButton
call NOTREACHED().

Bug: 933384
Change-Id: I17631d6e0dc1992c6cb2e38026584c224cf981cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1517034
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Peter Boström <pbos@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#640457}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 75a29468029806f5ad8c4b93fa4c2932a6f58006