Switch android-go-perf to use android-chrome.
This changes the android-go-perf bots on the main waterfall to start
using the android-chrome binary instead of android-chromium.

There are likely to be performance changes from this, but this is to
be expected.

Will send an announcement out before landing.

Bug: 931335
Change-Id: I68e6f61932331be58fe273bccb0885789feeb843
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1509194
Reviewed-by: Juan Antonio Navarro Pérez <perezju@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Simon Hatch <simonhatch@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#640685}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a55273847fd81449b903961093c5c78e58e8501f