Added new UMA to track the entirety of tab switching duration.
Created a new class TabSwitchTimeRecorder that tracks tab switching
duration from the input event timestamp until the frame is presented
on the screen. The the data will be uploaded on an UMA called
Browser.Tabs.TotalSwitchDuration.(WithSavedFrames|NoSavedFrames),
depending on whether we have saved frames when we switch to the tab.

Also merged the MPArch.RWH_TabSwitchPaintDuration metric tracking with
this new class, and removed the old callback creator for the
MPArch.RWH_TabSwitchPaintDuration metric.

Combining the MPArch.RWH_TabSwitchPaintDuration metric with this class
addresses the issue with that metric where sometimes the callback is
not fired on a tab switch, uploading wrong metric. See crbug/936858.

With this new UMA, we have data for the different stages on tab switch:
1. From input event timestamp to when the TabStripModel starts processing
   it (Browser.Tabs.InputEventToSelectionTime)
2. From when the TabStripModel starts processing tab switch until when
   the visibility request is sent by the Browser to the Renderer
   (Browser.Tabs.SelectionToVisibilityRequestTime)
3. From when the visibility request is sent until the frame presentation
   time (MPArch.RWH_TabSwitchPaintDuration)
4. The overall/total time (Browser.Tabs.TotalSwitchDuration)

Bug:921120

Change-Id: I72edae19bd885f4bae0bc9aab6e78bb221422a2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1494935
Commit-Queue: Edwin Joe <ejoe@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#640883}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a4a4c31a74c3541075a70ed60a984bf28cac4d34