Add UseCounter measurement to IdleDetection query
This change adds a use counter to measure the number of page loads which
use the IdleDetector feature. In another patch this will be enabled for
UKM so that we can follow-up in a year and see what popular sites
are using this.

Bug: 937896
Change-Id: If9d0f111c96da4cbb19e109d19e8a5c7823e6a83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1512541
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#642266}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8fe2be86be89ece76a527bc14ad409046d919179