Adding missing build dependency in search_engines
Generated code used a header the build target doesn't depend
on. This extracts that file to its own target so that
the "prepopulated_engines" can depend on that target.

This was found with an updated gn that also checks dependencies
for generated files (an understandable limitation in previous gn).

Bug: 655123
Change-Id: I4bee2a906c2e5c925db549e6bfa0be971afa3828
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1527535
Commit-Queue: Daniel Bratell <bratell@opera.com>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#642620}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a6a13a3ca04ff53a4eb59b3b4d573c904c8d9947