Fix Media.HardwareKeyPressed on CrOS
This histogram is not working on Chrome OS. Moves the enum
out of //content into //ui so both //content and //ash
can depend on it without having it duplicated.

Also adds a "PlayPause" action for when we do not know the
playback state.

BUG=942633

Change-Id: I3e8c895a60edd660cbf3c94426db706fa54b7999
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1526315
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: Becca Hughes <beccahughes@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#643051}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48530c07362c60254cbc527b8815500b87ce4cfb